[wip] typescript + refactoring
