--- Config/config.js
@@ -1,5 +1,0 @@
-const keys = {
-    MAPS_API_KEY: "AIzaSyDXb6dl90e3SsprhDQRxzuLm-oBRT2-khM",
-}
-
-export { keys }
