+++ Config/config.ts
@@ -1,0 +1,5 @@
+const keys = {
+    MAPS_API_KEY: 'AIzaSyDXb6dl90e3SsprhDQRxzuLm-oBRT2-khM'
+};
+
+export {keys};
