+++ Config/index.ts
@@ -2,8 +2,4 @@ import * as theme from './theme';
 import * as mocks from './mocks';
 import * as config from './config';
 
-export {
-  theme,
-  mocks, 
-  config, 
-};
+export {theme, mocks, config};
