+++ Config/mocks.ts
@@ -1,0 +1,26 @@
+const categories = [];
+
+const products = [];
+
+const explore = [
+    // images
+    require('../assets/explore_1.png'),
+    require('../assets/explore_2.png'),
+    require('../assets/explore_3.png'),
+    require('../assets/explore_4.png'),
+    require('../assets/explore_5.png'),
+    require('../assets/explore_6.png')
+];
+
+const profile = {
+    username: 'Max_Mustermann',
+    location: 'Europe',
+    email: 'Max_Mustermann@example.com',
+    avatar: require('../assets/avatar.png'),
+    budget: 1000,
+    monthly_cap: 5000,
+    notifications: true,
+    newsletter: false
+};
+
+export {categories, explore, products, profile};
