--- Helpers/googleMaps.js
@@ -1,6 +1,0 @@
-import { config } from '../Config'
-
-
-export function getAutoCompleteURL(query) {
-    let url = + config.keys.MAPS_API_KEY
-}
