+++ Helpers/googleMaps.ts
@@ -1,0 +1,6 @@
+import {config} from '../Config';
+
+// TODO: what does this even do?
+export function getAutoCompleteURL(query) {
+    let url = +config.keys.MAPS_API_KEY;
+}
