--- Resources/index.js
@@ -1,11 +1,0 @@
-const images = {
-    logo: require('../assets/logo.png')
-}
-
-const icons  = {
-    browseIcon : require ('../assets/browse.png'),
-    profileIcon : require ('../assets/user.png'),
-    addIcon : require ('../assets/add.png'),
-}
-
-export {images , icons}
