+++ Screens/AddDrive.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import {Text, View} from 'react-native';
+import {Block} from '../Components';
+
+export default function AddDrive() {
+    return (
+        <Block center middle>
+            <Text h1> Hier werden Fahrten hinzugefügt </Text>
+        </Block>
+    );
+}
