--- Screens/Filter.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react'
-import { Text, View } from 'react-native'
-import { Block } from '../Components'
-
-export default class Filter extends Component {
-    render() {
-        return (
-            <Block>
-                <Text> textInComponent </Text>
-            </Block>
-        )
-    }
-}
