+++ Screens/Filter.tsx
@@ -1,0 +1,11 @@
+import React, {Component} from 'react';
+import {Text, View} from 'react-native';
+import {Block} from '../Components';
+
+export default function Filter() {
+    return (
+        <Block>
+            <Text> textInComponent </Text>
+        </Block>
+    );
+}
