+++ Store/storeConfigurator.ts
