+++ package.json
@@ -44,7 +44,8 @@
     "redux": "^4.0.5"
   },
   "devDependencies": {
-    "@babel/core": "~7.13.0"
+    "@babel/core": "~7.13.0",
+    "@types/react-native": "0.64.3"
   },
   "private": true
 }
