+++ tsconfig.json
@@ -1,0 +1,7 @@
+{
+    "compilerOptions": {
+        "forceConsistentCasingInFileNames": true,
+        "strict": true
+    },
+    "extends": "expo/tsconfig.base"
+}
